week1/2-labs/02-Model_with_fastapi/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
pillow
streamlit
joblib